prototype: -> ProtectedRoute.ts
