+++ react_type_script/src/main/webapp/app/common/types/server-api-dtos.ts
@@ -11,7 +11,7 @@ export type ResponseJSON = {
   responseJSON: ErrorDto;
 };
 
-enum UserRoleEnum {
+export enum UserRoleEnum {
   ADMIN = 'ADMIN',
   CHIEF = 'CHIEF',
   OPERATOR = 'OPERATOR',
