--- react_type_script/src/main/webapp/app/components/protected-route/utils.js
@@ -1,4 +1,0 @@
-/*
- * Copyright (c) 2020. Prototype
- */
-
