+++ react_type_script/src/main/webapp/package.json
@@ -107,6 +107,7 @@
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.7.4",
     "@types/enzyme": "^3.1.15",
+    "@types/react-router-dom": "^5.1.3",
     "@typescript-eslint/eslint-plugin": "^2.11.0",
     "@typescript-eslint/parser": "^2.11.0",
     "autoprefixer": "9.1.1",
